NLA-SLR/requirements.txt
# azureml==0.2.7
# azureml_core==1.44.0
Lintel==1.0
matplotlib==3.1.2
numpy==1.21.6
opencv-python==4.2.0.32
Pillow==9.2.0
PyYAML==6.0
scikit_learn==1.1.2
timm==0.4.12
torch==1.9.0+cu102
torchvision==0.10.0
tqdm==4.62.3
transformers==4.11.3
wandb==0.12.17